refactor: reRoute・lowerCamel
