+++ src/BlogCard.ts
@@ -1,5 +1,5 @@
 import { TagListItem } from './TagList';
-import { formatYMD } from './lib/date-utils';
+import { formatYMD } from './lib/dateUtils';
 import { useMediaQueryContext } from './mediaQueryContext';
 import { Post } from './post';
