+++ src/lib/dateConstants.ts
@@ -1,4 +1,4 @@
-const second = 1000;
+const second = 1000; // ms
 const minute = second * 60;
 const hour = minute * 60;
 const day = hour * 24;
